in_tail: Rename group_watcher.rb to group_watch.rb
