+++ lib/fluent/plugin/in_tail/group_watch.rb
